20230227-done-spring-#3: 섹션 06. 합성 애노테이션을 활용한 스프링 부트 만들기
